requirements.txt
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
python-dotenv==1.0.0
alembic==1.12.1
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
aiokafka==0.10.0
httpx==0.25.2
